feat(Profile and ReservationUser): split page UserProfile into 2 new pages Profile and ReservationUser
