+++ src/pages/ReservationUser/ReservationUser.styles.ts
@@ -1,35 +1,20 @@
 import styled from 'styled-components';
 
-const UpdateButtonContainer = styled.div`
+const ToggleButtonsContainer = styled.div`
+    max-width: 93.75rem;
     display: flex;
-    flex-direction: column;
+    gap: 1.875rem;
+    margin: 5rem auto 1.875rem;
+    padding: 0 2rem;
+    overflow: hidden;
     align-items: center;
     justify-content: center;
-    width: 100%;
-    height: 100%;
-    padding: 0.625rem;
-    gap: 0.625rem;
-`;
+    border-radius: 0.5rem;
 
-const StyledDetailBullet = styled.li`
-    display: flex;
-    align-items: center;
-    margin: 0.5rem 1.25rem;
-    background-color: var(--beige-light);
-    padding: 0.5rem;
-    border-radius: 0.25rem;
-    overflow: hidden;
-    img {
-        margin-top: 0.1875rem;
-        margin-right: 0.3125rem;
-        width: 2.5rem;
-        height: 2.5rem;
-    }
-
-    span {
-        margin-bottom: 0.1875rem;
-        font-size: 1rem;
-        font-weight: bold;
+    @media (max-width: 768px) {
+        margin: 1.25rem auto 1.875rem;
+        flex-direction: column;
+        gap: 0.3125rem;
     }
 `;
 
@@ -56,6 +41,18 @@ const UserProfileContainer = styled.div`
     }
 `;
 
+const UserAdditionalInfoContainer = styled.div`
+    display: flex;
+    flex-direction: column;
+    margin: 0.625rem auto;
+    width: 90%;
+    height: 90%;
+    background-color: var(--brown);
+    border-radius: 0.5rem;
+    text-align: center;
+    justify-content: center;
+`;
+
 const UserMainInfoContainer = styled.div`
     display: flex;
     flex-direction: row;
@@ -72,31 +69,6 @@ const UserMainInfoContainer = styled.div`
     @media (max-width: 510px) {
         flex-direction: column;
     }
-`;
-const UserProfilePictureContainer = styled.div`
-    display: flex;
-    flex-direction: column;
-    margin: 0.625rem auto;
-    width: 90%;
-    height: 90%;
-
-    border-radius: 0.5rem;
-    align-items: center;
-    justify-content: center;
-    text-align: center;
-`;
-
-const UserProfilePicture = styled.img`
-    border-radius: 50%;
-    width: 12.5rem;
-    height: 12.5rem;
-    object-fit: cover;
-    border: 0.1875rem solid var(--blue-green-light);
-    margin-bottom: 0.625rem;
-`;
-
-const UserProfileEmail = styled.p`
-    font-size: 1.2rem;
 `;
 
 const ListContainer = styled.div`
@@ -129,13 +101,10 @@ const BackButton = styled.button`
 `;
 
 export {
-    UpdateButtonContainer,
-    StyledDetailBullet,
+    ToggleButtonsContainer,
+    UserAdditionalInfoContainer,
     UserProfileContainer,
     UserMainInfoContainer,
-    UserProfilePicture,
-    UserProfilePictureContainer,
-    UserProfileEmail,
     ListContainer,
     BackButton,
 };
