+++ src/static/routes.ts
@@ -12,6 +12,7 @@ export const route = {
     updateFloorPlan: ':fpId/update',
     createReservation: '/create-reservation',
     reservationSummary: '/reservation-summary',
+    reservationUser: 'reservation/user',
     calendar: '/calendar',
     user: '/user',
     updatePassword: '/user-update-password/:userId',
